# Try lower DB pool size for Crystal